Time: 2 ms (99.89%), Space: 56.2 MB (65.29%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -13,6 +13,6 @@ class Solution {
             windowSum+= nums[i] - nums[i-k];
             maxSum = Math.max(maxSum,windowSum);
         }
-        return (double)maxSum / (double)k;        
+        return (double)maxSum /k;        
     }
 }
